+++ lib/resource/template/textfield/TextFieldTemplateBase.dart
@@ -114,7 +114,6 @@ class TextFieldTemplateBase    {
     double widthUnderLine = 1;
     if( isRemoveUnderline ) {
       widthUnderLine = 0;
-      hint_color = Colors.transparent;
     }
 
     var enabledBorder = UnderlineInputBorder(